Rename code modules as resources
