+++ src/aiidalab_qe/plugins/bands/resources.py
