+++ src/aiidalab_qe/plugins/pdos/__init__.py
@@ -1,7 +1,7 @@
 from aiidalab_qe.common.panel import PluginOutline
 
-from .code import PdosResourceSettingsModel, PdosResourceSettingsPanel
 from .model import PdosConfigurationSettingsModel
+from .resources import PdosResourceSettingsModel, PdosResourceSettingsPanel
 from .result import PdosResultsModel, PdosResultsPanel
 from .setting import PdosConfigurationSettingPanel
 from .workchain import workchain_and_builder
