+++ src/aiidalab_qe/plugins/xas/resources.py
